Remove take() & skip() constructs
I did this because I couldn't get it to work with the behaviors I wanted
when dealing with taking from a Void parser vs taking from a non-Void
parser. In the Void case we would just want to return the new parsers
Output. In the non-Void case we would want to return a tuple of the
non-Void parsers output and the new parsers Output.

But I couldn't figure out a way to properly setup type constraints to
facilitate this. Also, I think take & skip make it less clear how the
composition of parsers works.

Also it is harder to reason about from a maintenance standpoint.

So I decided to just remove take & skip from the library and just use
the functional operations like zip, map, flatMap, etc. for the
composition of parsers.

[changelog]
removed: take() & skip() interface from library

ps-id: 76ADD10D-5794-4C51-B35F-557976964DDD