Cleanup type constraints & add .first()
I cleaned up the type constraints by moving them from the Parser type to
the function for all the static functions. I also cleaned up the type
constraints on the few methods that are actually on the Parser type so
that both the Parser type constraints make sense as well as the function
type constraints.

In addition to that clean up I added the .first(p: Parser) static
function to facilitate reaching into a Collection and having a parser
run on the first element of the Collection while returning the result of
that inner parsing.

[changelog]
added: first() to facilitate element parsing of a collection
changed: type constraints of everything in the library

ps-id: 00A71108-2CDF-45A9-BBEB-3527B3695C53