Fix bug with Collection.range(of:)
There was a bug with our implementation of range(of:) on Collection
where it was missing the last element in it's checks. This corrects that
off by one error so that it can now properly match up to the last
element.

[changelog]
fixed: bug causing parsers to not match up the last element

ps-id: 2CF1ED89-7E7C-4375-B93B-BC11133B0C1C